feat(rocket): add auth guard for rocket oauth2.0 introspection (#319)
This adds rocket as a possible feature to have a
OAuth2.0 Token introspection route guard.
The guard does check the incoming token
and introspects it against a given ZITADEL
instance. A new feature "rocket" is used
to enable all rocket specific stuff.

BREAKING CHANGE: This removes `api` as a default
feature. To migrate, just add `api` to the used features
of the crate.