feat(clients): Add v2 API clients and centralize client generation (#522)
This allows the new v2 APIs of ZITADEL to be used.

BREAKING CHANGE: The modules for client instantiation
have changed. There is no module "with access token"
or "with service account" anymore. Instead, there is a
centralized `ClientBuilder` in the clients module of the
api to create and configure the gRPC clients. The new
builder allows configuring auth methods beforehand
and then create the according client out of it.