feat(introspection_cache): remove custom role generic to make compatible with caches
Using a role generic was incompatible with the type erasure
of the introspection cache.
Judging from the discussion in https://github.com/smartive/zitadel-rust/pull/550
the original intention was to introduce a convenience feature, for library users.
However, the same behaviour can be easily achieved by a simple extention trait,
and does not need to be support by this crate itself to achieve the same convenience.
As a result, the generic role type parameter is removed.