fix: log error responses returned by Zitadel
When we receive an error response on an introspection call, for
example a 503, we previously returned a ParseError as the response
was not valid JSON. This change adds a new error type
ZitadelResponseError which handles such cases by logging the
status code and body of the error response. This is a breaking
change as we add a Variant to the error enum which needs
to be handled by the user.